Update to 1.1.0 - Added more support for ansible_group, and metadatas.
